Pile of minor fixes
- Fix initializing things inside switch case in "Render" fn for icosphere
- Fix not including Descriptor types in vtfFrameData
- Fix include paths that were using signal/Delegate.hpp
- Remove dated functions in Rendergraph.cpp
- Remove completely broken destructor w missing declaration in DescriptorTemplate
- Fix a number of casting errors relating to size_t/uint32_t collisions and issues
- Remove duplicated function declaration in UpdateTemplateData
- Use uint32_t more as stored type in classes, since Vulkan uses it as it's numeric type so often
- Fix using function before it was declared/defined in vtfTasksAndSteps.cpp

+++ modules/rendergraph_module/include/Descriptor.hpp
@@ -60,12 +60,12 @@ private:
     void allocateSets();
     void createPool();
 
-    size_t maxSets{ 0u };
+    uint32_t maxSets{ 0u };
     const vpr::Device* device{ nullptr };
     DescriptorTemplate* templ{ nullptr };
     std::stack<std::unique_ptr<vpr::DescriptorPool>> descriptorPools;
     vpr::DescriptorPool* activePool{ nullptr };
-    std::atomic<size_t> setContainerIdx;
+    std::atomic<uint32_t> setContainerIdx;
     std::vector<VkDescriptorSet> availSets;
     std::stack<std::vector<VkDescriptorSet>> usedSets;
     st::descriptor_type_counts_t typeCounts;
